requirements.txt
einops==0.8.1
numpy==1.24.3
pandas==2.1.3
Pillow==11.2.1
scikit-learn==1.3.0
scipy==1.10.1
tensorboardX==2.6.2.2
timm==1.0.15
torch==2.4.1
torchvision==0.15.2a0
tqdm==4.66.4